renamed contract
