+++ contracts/snip20_staking/Cargo.toml
@@ -45,7 +45,7 @@ rand_chacha = { version = "0.2.2", default-features = false }
 rand_core = { version = "0.5.1", default-features = false }
 sha2 = { version = "0.9.1", default-features = false }
 
-shade-protocol = { version = "0.1.0", path = "../../packages/shade_protocol", features = ["shd_staking", "snip20", "storage"] }
+shade-protocol = { version = "0.1.0", path = "../../packages/shade_protocol", features = ["snip20_staking", "snip20", "storage"] }
 cosmwasm-math-compat = { version = "0.1.0", path = "../../packages/cosmwasm_math_compat" }
 
 # Large number math
