+++ packages/shade_protocol/Cargo.toml
@@ -35,7 +35,7 @@ mint = ["utils", "snip20"]
 mint_router = ["utils", "snip20"]
 oracle = ["snip20"]
 scrt_staking = ["utils"]
-shd_staking = ["utils"]
+snip20_staking = ["utils"]
 treasury = ["utils"]
 
 # for quicker tests, cargo test --lib
